Реализовал метод удаления-DeleteView

+++ api_v2/views.py
@@ -61,4 +61,8 @@ class ArticleView(APIView):
     def delete(self, request, *args, **kwargs):
         pk = self.kwargs.get('pk')
         article = get_object_or_404(Article, pk=pk)
-        article.delete()
+        if pk:
+            article.delete()
+            return Response({'deleted_id': pk})
+        else:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
